cli: export PrepareEnvironment
Other tools that consume the ESC CLI can use this method to extract
environment variables, create temporary files, and discover associated
secrets from an environment.
