+++ cmd/esc/cli/env_run.go
@@ -148,7 +148,7 @@ func newEnvRunCmd(envcmd *envCommand) *cobra.Command {
 				return envcmd.writePropertyEnvironmentDiagnostics(envcmd.esc.stderr, diags)
 			}
 
-			files, environ, secrets, err := envcmd.prepareEnvironment(env, prepareOptions{})
+			files, environ, secrets, err := envcmd.prepareEnvironment(env, PrepareOptions{})
 			if err != nil {
 				return err
 			}
